fix: clean up code
I lowered the minimum length name after noticing it was too large (and
that I couldn't put my own name in). I also placed it in a constant
variable MIN_LENGTH to avoid magic numbers.

Firefox was complaining about it not being able to detect the difference
between gameStarted and gamestarted, so changing the variable in <room>
to game-started and wrapping the app's variable gameStarted in quotation
marks resolved that.

After noticing that the styling was the same for p and h1,h2,h3, I
deleted the p's styling and appended p to h1,h1,h3. I also added li to
the styling rules.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>