fix: cover edge case errors in room join
Attempting to join the game while it is in play no longer works. I added
a condition to check if the game was in play when requesting to join the
room.

Additionally, checking if the username was already in use did not work
as intended. They would still enter the room, since there was no return
statement. Now that the return statement is there, username duplication
checking rejects as intended.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>